requirements.txt
numpy
PyQt6
pillow
torch
